requirements.txt
cloudinary
Django==3.2
djangorestframework
djangorestframework-simplejwt